+++ bitcoin/src/lib.rs
@@ -92,6 +92,7 @@ mod serde_utils;
 pub mod network;
 pub mod address;
 pub mod amount;
+pub mod base58;
 pub mod bip152;
 pub mod bip158;
 pub mod bip32;
